+++ app/src/main/java/de/triology/cas/services/RegistryLocal.java
@@ -19,10 +19,12 @@ import java.util.*;
 import java.util.stream.Collectors;
 
 @Slf4j
-public class RegistryLocal implements Registry{
+public class RegistryLocal implements Registry {
 
-    private static final String LOCAL_CONFIG_FILE = "/var/ces/config/local.yaml";
     private static final String LOCAL_CONFIG_DIR = "/var/ces/config";
+    private static final String LOCAL_CONFIG_SUB_PATH = "/var/ces/config";
+    private static final String LOCAL_CONFIG_FILE_NAME = "local.yaml";
+    private static final String LOCAL_CONFIG_FILE = LOCAL_CONFIG_SUB_PATH + "/" + LOCAL_CONFIG_FILE_NAME;
     private static final String GLOBAL_CONFIG_FILE = "/etc/ces/config/global/config.yaml";
 
     FileSystem fileSystem = FileSystems.getDefault();
@@ -32,11 +34,13 @@ public class RegistryLocal implements Registry{
     protected static class GlobalConfig {
         private String fqdn;
     }
+
     @Getter
     @Setter
     protected static class LocalConfig {
         private ServiceAccounts service_accounts = new ServiceAccounts();
     }
+
     @Getter
     @Setter
     protected static class ServiceAccountSecret {
@@ -48,6 +52,7 @@ public class RegistryLocal implements Registry{
             return stringsEqual(secret, other.secret) && stringsEqual(logout_uri, other.logout_uri);
         }
     }
+
     @Getter
     @Setter
     protected static class ServiceAccountCas {
@@ -217,31 +222,69 @@ public class RegistryLocal implements Registry{
 
     @Override
     public void addDoguChangeListener(DoguChangeListener doguChangeListener) {
-        Thread t1 = new Thread(() -> {
-            var path = Paths.get(LOCAL_CONFIG_DIR);
-            try(WatchService watchService = fileSystem.newWatchService()) {
-                path.register(watchService, StandardWatchEventKinds.ENTRY_CREATE, StandardWatchEventKinds.ENTRY_MODIFY);
-                var previousServiceAccounts = readServiceAccounts();
-                while (true) {
-                    LOGGER.info("wait for changes under {}", LOCAL_CONFIG_DIR);
-                    watchService.take();
+        Thread t1 = new Thread(() -> doguChangeListenerThreadHandler(doguChangeListener));
+        t1.start();
+    }
+
+    private void doguChangeListenerThreadHandler(DoguChangeListener doguChangeListener) {
+        var path = Paths.get(LOCAL_CONFIG_DIR);
+        WatchService watchService;
+        try {
+            watchService = initializeWatchService(path, null);
+            var previousServiceAccounts = readServiceAccounts();
+            WatchKey key;
+            LOGGER.info("wait for changes under {}", LOCAL_CONFIG_DIR);
+            while ((key = watchService.take()) != null) { // take does not return null
+                if (!key.isValid()) {
+                    LOGGER.info("watch key was cancelled or watch service was closed");
+                    watchService = initializeWatchService(path, watchService);
+                    continue;
+                }
+
+                List<WatchEvent<?>> watchEvents = key.pollEvents();
+                boolean serviceAccountRegistryChanged = false;
+                for (WatchEvent<?> event : watchEvents) {
+                    if (event.context() != null && event.context().toString().equals(LOCAL_CONFIG_FILE_NAME)) {
+                        serviceAccountRegistryChanged = true;
+                        break;
+                    }
+                }
+
+                if (serviceAccountRegistryChanged) {
                     var currentServiceAccounts = readServiceAccounts();
                     if (!previousServiceAccounts.deepEquals(currentServiceAccounts)) {
+                        LOGGER.info("services changed. call doguChangeListener");
                         doguChangeListener.onChange();
                         previousServiceAccounts = currentServiceAccounts;
                     }
                 }
-            } catch (IOException e) {
-                throw new RegistryException("Failed to addDoguChangeListener", e);
-            } catch (InterruptedException e) {
-                Thread.currentThread().interrupt();
-                throw new RegistryException("Failed to addDoguChangeListener", e);
+
+                // Resetting the watchKey is very important. Otherwise, the key returned from take() (or poll()) will not return any more events.
+                if (!key.reset()) {
+                    LOGGER.info("watch key is no longer valid");
+                    watchService = initializeWatchService(path, watchService);
+                }
             }
-        });
-        t1.start();
+        } catch (IOException e) {
+            throw new RegistryException("Failed to addDoguChangeListener", e);
+        } catch (InterruptedException e) {
+            Thread.currentThread().interrupt();
+            throw new RegistryException("Failed to addDoguChangeListener", e);
+        }
     }
 
-    private static <T> T readYaml(Class<T> tClass, InputStream yamlStream){
+    private WatchService initializeWatchService(Path path, WatchService oldWatchService) throws IOException {
+        if (oldWatchService != null) {
+            LOGGER.info("close old watch service");
+            oldWatchService.close();
+        }
+        LOGGER.info("initialize watch service");
+        WatchService watchService = fileSystem.newWatchService();
+        path.register(watchService, StandardWatchEventKinds.ENTRY_CREATE, StandardWatchEventKinds.ENTRY_MODIFY);
+
+        return watchService;
+    }
+    private static <T> T readYaml(Class<T> tClass, InputStream yamlStream) {
         var yaml = new Yaml(new Constructor(tClass, new LoaderOptions()));
         try {
             T result = yaml.load(yamlStream);
@@ -254,9 +297,9 @@ public class RegistryLocal implements Registry{
         } catch (YAMLException e) {
             throw new RegistryException(String.format("Failed to parse yaml stream to class %s", tClass.getName()), e);
         } catch (InstantiationException |
-                IllegalAccessException |
-                InvocationTargetException |
-                NoSuchMethodException e) {
+                 IllegalAccessException |
+                 InvocationTargetException |
+                 NoSuchMethodException e) {
             throw new RegistryException(String.format("Failed to construct new instance of %s", tClass.getName()), e);
         }
     }