+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v7.0.8-2] - 2024-10-02
+### Fixed
+- Fix a bug where the watch for service accounts in the config `local.yaml` stucks because the events wasn't resetted and polled [#217]
+
 ## [v7.0.8-1] - 2024-09-23
 ### Added
 - Add "lang"-attribute to HTML-Pages [#213]